get-file: make root an optional kwarg; patch url formatting.
This latter tweak needs a bit of explanation.  The path argument was
spliced into the url winding up along the lines of
"..files/auto/<path>", meaning that if you took the :path entry from a
/delta response, which is fully-qualified, the url would be
"..files/auto//filename", and cause an error.

Additionally, the :root parameter is recommended to be "auto", so we
just make that the default and don't require it.